+++ sdk/src/rhesis/sdk/metrics/tests/test_rhesis.py
@@ -4,8 +4,8 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from rhesis.sdk.metrics.base import MetricResult
-from rhesis.sdk.metrics.rhesis import RhesisPromptMetric
-from rhesis.sdk.metrics.rhesis.metric_base import RhesisMetricBase, ScoreType, ThresholdOperator
+from rhesis.sdk.metrics.providers.rhesis import RhesisPromptMetric
+from rhesis.sdk.metrics.providers.rhesis import RhesisMetricBase, ScoreType, ThresholdOperator
 
 
 @pytest.fixture
